tests: avoid definitions in headers
symbols not exported on Windows
